requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
psycopg2-binary==2.9.5
pydantic==1.10.2